Modules/requirements.txt
streamlit==1.28.0
psycopg2-binary==2.9.7
pandas==2.1.0
plotly==5.15.0
sqlalchemy==2.0.19
openai==1.3.0
python-dotenv==1.0.0
mcp-server-postgres==0.1.0
uvicorn==0.23.0
fastapi==0.103.0
asyncio-mqtt==0.13.0